# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Notebook API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PasswordInput {
  password: string;
}

export interface AuthResult {
  success: boolean;
}

export interface UploadResult {
  url: string;
}

export interface Piece {
  id: number;
  title: string;
  composer: string;
  composerId?: number | null;
  year?: string | null;
  genre?: string | null;
  youtubeUrl?: string | null;
  imageUrl?: string | null;
  content?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface PieceInput {
  title: string;
  composer: string;
  composerId?: number | null;
  year?: string | null;
  genre?: string | null;
  youtubeUrl?: string | null;
  imageUrl?: string | null;
  content?: string | null;
  tags?: string[];
}

export interface Composer {
  id: number;
  name: string;
  born?: string | null;
  died?: string | null;
  nationality?: string | null;
  imageUrl?: string | null;
  biography?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ComposerInput {
  name: string;
  born?: string | null;
  died?: string | null;
  nationality?: string | null;
  imageUrl?: string | null;
  biography?: string | null;
}

export interface Note {
  id: number;
  title: string;
  section: string;
  chapterTitle?: string | null;
  imageUrl?: string | null;
  content?: string | null;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  title: string;
  section: string;
  chapterTitle?: string | null;
  imageUrl?: string | null;
  content?: string | null;
  sortOrder: number;
}

export interface Book {
  id: number;
  title: string;
  author?: string | null;
  coverUrl?: string | null;
  synopsis?: string | null;
  year?: string | null;
  content?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BookInput {
  title: string;
  author?: string | null;
  coverUrl?: string | null;
  synopsis?: string | null;
  year?: string | null;
  content?: string | null;
}

export type UploadImageBody = {
  file: Blob;
};