# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Notebook API
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  title: string;
  author?: string | null;
  coverUrl?: string | null;
  synopsis?: string | null;
  year?: string | null;
  content?: string | null;
}