# lib/api-zod/src/generated/types/composerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Notebook API
 * OpenAPI spec version: 0.1.0
 */

export interface ComposerInput {
  name: string;
  born?: string | null;
  died?: string | null;
  nationality?: string | null;
  imageUrl?: string | null;
  biography?: string | null;
}