# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Notebook API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteInput {
  title: string;
  section: string;
  chapterTitle?: string | null;
  imageUrl?: string | null;
  content?: string | null;
  sortOrder: number;
}