# lib/api-zod/src/generated/types/piece.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Notebook API
 * OpenAPI spec version: 0.1.0
 */

export interface Piece {
  id: number;
  title: string;
  composer: string;
  composerId?: number | null;
  year?: string | null;
  genre?: string | null;
  youtubeUrl?: string | null;
  imageUrl?: string | null;
  content?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}